Major refactor of world, dimension, and chunk
all are now abstract
