+++ chunky/src/java/se/llbit/chunky/ui/render/tabs/GeneralTab.java
@@ -17,6 +17,7 @@
  */
 package se.llbit.chunky.ui.render.tabs;
 
+import it.unimi.dsi.fastutil.ints.IntIntPair;
 import javafx.beans.binding.Bindings;
 import javafx.beans.value.ChangeListener;
 import javafx.fxml.FXML;
@@ -592,12 +593,14 @@ public class GeneralTab extends ScrollPane implements RenderControlsTab, Initial
   }
 
   private void updateYClipSlidersRanges(World world) {
-    if (world != null && world.getVersionId() >= World.VERSION_21W06A) {
-      yMin.setRange(-64, 320);
-      yMax.setRange(-64, 320);
-    } else {
-      yMin.setRange(0, 256);
-      yMax.setRange(0, 256);
+    if (world != null) {
+      IntIntPair heightRange = world.currentDimension().heightRange();
+      int min = heightRange.firstInt();
+      int max = heightRange.secondInt();
+      yMin.setRange(min, max);
+      yMin.set(min);
+      yMax.setRange(min, max);
+      yMax.set(max);
     }
   }
 }
