+++ chunky/src/java/se/llbit/chunky/world/region/MCRegion.java
@@ -24,6 +24,8 @@ import java.util.Set;
 import se.llbit.chunky.chunk.ChunkLoadingException;
 import se.llbit.chunky.plugin.PluginApi;
 import se.llbit.chunky.world.*;
+import se.llbit.chunky.world.java.JavaChunk;
+import se.llbit.chunky.world.java.JavaDimension;
 import se.llbit.log.Log;
 import se.llbit.nbt.ErrorTag;
 import se.llbit.nbt.NamedTag;
@@ -63,7 +65,7 @@ public class MCRegion implements Region {
 
   private final Chunk[] chunks = new Chunk[NUM_CHUNKS];
   private final RegionPosition position;
-  private final Dimension dimension;
+  private final JavaDimension dimension;
   private final String fileName;
   private long regionFileTime = 0;
   private final int[] chunkTimestamps = new int[NUM_CHUNKS];
@@ -80,7 +82,7 @@ public class MCRegion implements Region {
    *
    * @param pos the region position
    */
-  public MCRegion(RegionPosition pos, Dimension dimension) {
+  public MCRegion(RegionPosition pos, JavaDimension dimension) {
     this.dimension = dimension;
     fileName = pos.getMcaName();
     position = pos;
@@ -156,7 +158,7 @@ public class MCRegion implements Region {
           int loc = file.readInt();
           if (loc != 0) {
             if (chunk.isEmpty()) {
-              chunk = new Chunk(pos, dimension);
+              chunk = new JavaChunk(pos, dimension);
               setChunk(pos, chunk);
             }
           } else {
